% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.R
\name{a}
\alias{a}
\title{Shorthand for \code{as.data.frame()} for printing all of a tibble.}
\usage{
a(tib)
}
\arguments{
\item{tib}{\code{tibble} to print}
}
\value{
print \code{a} as a data.frame, so show all the rows
}
\description{
Shorthand for \code{as.data.frame()} for printing all of a tibble.
}
\details{
Often want to print all of a (small) tibble, but \code{as.data.frame()}
is cumbersome.
}
\author{
Andrew Edwards
}
