% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_w_to_b_l.R
\name{b_l_to_b_w}
\alias{b_l_to_b_w}
\title{Converting between length- and weight-based size spectrum exponents
Will write out the math in a write up, but the equations are:
b_l = \beta (b_w + 1) - 1
b_w = (b_l + 1) / \beta - 1}
\usage{
b_l_to_b_w(b_l, beta = 3.1802)
}
\arguments{
\item{b_l}{exponent for a length-based individual size spectrum}

\item{beta}{exponent of the length-weight relationship for the species in}
}
\value{
the \code{b_w} exponent for the weight-based individual size spectrum
}
\description{
Converting between length- and weight-based size spectrum exponents
Will write out the math in a write up, but the equations are:
b_l = \beta (b_w + 1) - 1
b_w = (b_l + 1) / \beta - 1
}
\examples{
\dontrun{
b_weight <- b_l_to_b_w(-5.1, beta = 3)
}
}
\author{
Andrew Edwards
}
