% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts_per_bin_example}
\alias{counts_per_bin_example}
\title{Example tibble in the right format to use in testing and examples,}
\format{
Tibble with self-explanatory columns:
\describe{
\item{binCount}
\item{binMid}
\item{binMin}
\item{binMax}
}
}
\source{
Generated from running \code{data-raw/counts_per_bin_example.R}.
}
\usage{
counts_per_bin_example
}
\description{
Has a gap (empty bins that aren't explicitly mentioned) before the final bin
for testing \code{make_hist()}, is loosely based on 2004 data.
}
\examples{
\dontrun{
counts_per_bin_example
make_hist(counts_per_bin_example)
}

}
\author{
Andrew Edwards
}
\keyword{datasets}
