% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_all_years.R
\name{fit_all_years}
\alias{fit_all_years}
\title{Fit all the years of data using MLEbin method}
\usage{
fit_all_years(
  raw_simp_prop,
  strata = c("C", "NC", "S", "SC", "N"),
  min_year_to_analyse,
  max_year_to_analyse = NULL,
  bin_width_each_year
)
}
\arguments{
\item{raw_simp_prop}{input tibble of correct format (see .Rmd)}

\item{strata}{which strata to use, default is the full coast (excluding ones
that should not be used)}

\item{min_year_to_analyse}{minimum year to analyse}

\item{max_year_to_analyse}{maximum year to analyse, if \code{NULL} (the default)
then does the last year of data}

\item{bin_width_each_year}{tibble of bin widths in each year, calculated
earlier TODO could include as a data object in the package}
}
\value{
list of class \code{hake_spectra_results} with one element for each year,
each one containing:
\describe{
\item{year}
\item{bin_width}
\item{xmin}
\item{xmax}
\item{n}
\item{counts_per_bin}
\item{counts_per_bin_desc}{counts in the descending limb, including the
peak}
\item{b_l}
\item{b_l_confMin}
\item{b_l_confMax}
}
}
\description{
Fit all the years of data using MLEbin method
}
\examples{
\dontrun{
# see .Rmd file
res_all_strata <- fit_all_years(raw_simp_prop,
                    bin_width_each_year = bin_width_each_year)
}
}
\author{
Andrew Edwards
}
