% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLEbin_results_extract.R
\name{MLEbin_results_extract}
\alias{MLEbin_results_extract}
\title{Extract MLEbin results for each year from a \code{res_all_years} list}
\usage{
MLEbin_results_extract(res_all_years)
}
\arguments{
\item{res_all_years}{list object from \code{fit_all_years()}}
}
\value{
tibble of MLEbin results
}
\description{
Extract MLEbin results for each year from a \code{res_all_years} list
}
\examples{
\dontrun{
MLEbin_results_extract(res_all_years) # See .Rmd
}
}
\author{
Andrew Edwards
}
