% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.hake_spectra_results.R
\name{plot.hake_spectra_results}
\alias{plot.hake_spectra_results}
\title{Plot hake spectra results}
\usage{
\method{plot}{hake_spectra_results}(res, years = NULL, ...)
}
\arguments{
\item{res}{list of class \code{hake_spectra_results} as output from \code{fit_all_years()}}

\item{years}{years to be plotted (if \code{NULL} then do all years)}

\item{...}{arguments to passed onto \code{sizeSpectra::ISD_bin_plot_nonoverlapping()}}
}
\value{
figure in current device, see .Rmd for multiple pages
}
\description{
Plot hake spectra results
}
\examples{
\dontrun{
See .Rmd
}
}
\author{
Andrew Edwards
}
