% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_b_recruitment.R
\name{plot_b_recruitment}
\alias{plot_b_recruitment}
\title{Plot estimates of b and recruitment}
\usage{
plot_b_recruitment(
  b_tibble,
  recruitment_tibble = hake_recruitment_over_2010,
  years_exclude = NULL,
  conf_col = "grey",
  label_col = "red",
  xlim = NULL,
  ylim = c(0, 1.1),
  xlab = "Length size spectrum exponent, b_l",
  ylab = "Estimated age-0 hake recruitment scaled by 2010 value",
  pch = 20,
  cex = 1,
  text_cex = 0.7,
  adj = c(-0.15, -0.3),
  show_prior = TRUE,
  prior_loc = 0.3,
  ...
)
}
\arguments{
\item{b_tibble}{currently with \code{Year}, \code{b_l}, \code{b_l_confMin}, \code{b_l_confMax},}

\item{recruitment_tibble}{tibble with same years as \code{b_tibble}, with columns
\code{year}, \code{low}, \code{median}, \code{high}}

\item{years_exclude}{vector of years to exclude the \code{b_tibble} results for; if NULL then
use all years}

\item{conf_col}{colour for confidence intervals}

\item{label_col}{colour for labels}

\item{xlim}{usual, default encompass all years being plotted}

\item{ylim}{usual default includes max of 2023 credible interval}

\item{xlab}{usual}

\item{ylab}{usual}

\item{pch}{usual}

\item{cex}{usual}

\item{text_cex}{for year labels}

\item{adj}{adjustment for labelling years}

\item{show_prior}{logical, whether to show prior distribution}

\item{prior_loc}{x-value to plot prior distribution}

\item{...}{extra arguments for \code{plot()} TODO not implemented yet}
}
\value{
plots recruitment estimates (with credible intervals) against \code{b_l}
estimates (with credible intervals)
}
\description{
Adds the prior, which is assumed to be the final year in \code{recruitment}
(which will be the prior for hake).
TODO Could add switch for \code{b_w}
TODO subscript x label properly
}
\examples{
\dontrun{
# See .Rmd
}
}
\author{
Andrew Edwards
}
