% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_l_to_b_w.R
\name{b_l_to_b_w}
\alias{b_l_to_b_w}
\alias{b_w_to_b_l}
\title{Converting between length- and weight-based size spectrum exponents}
\usage{
b_l_to_b_w(b_l, beta = 3.1802)

b_w_to_b_l(b_w, beta = 3.1802)
}
\arguments{
\item{b_l}{exponent for a length-based individual size spectrum}

\item{beta}{exponent of the length-weight relationship for the species in}

\item{b_w}{exponent for a weight-based individual size spectrum}
}
\value{
the \code{b_w} exponent for the weight-based individual size spectrum

the \code{b_l} exponent for the length-based individual size spectrum
}
\description{
Will write out the math in a write up, but the equations are:
\deqn{b_l = \beta (b_w + 1) - 1}
and
\deqn{b_w = (b_l + 1) / \beta - 1}
}
\examples{
\dontrun{
b_weight <- b_l_to_b_w(-5.1, beta = 3)
}
\dontrun{
b_length <- b_w_to_b_l(-2, beta = 3)
}
}
\author{
Andrew Edwards
}
